CORE-1282: split ClientInfo and ContactSupport into separate source files for reuse
