+++ src/components/utils/error/ErrorHandler.js
@@ -3,21 +3,19 @@
  *
  * A component that displays formatted error message with options to contact support or login
  */
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useRouter } from "next/router";
 import { useTranslation } from "i18n";
 
-import { intercomShow } from "../../../common/intercom";
 import NavigationConstants from "../../../common/NavigationConstants";
 import GridLabelValue from "../GridLabelValue";
-import { useUserProfile } from "../../../contexts/userProfile";
 import constants from "../../../constants";
-import GridLoading from "../GridLoading";
+import ClientInfo from "./ClientInfo";
+import ContactSupport from "./ContactSupport";
 
 import { build } from "@cyverse-de/ui-lib";
 import ids from "../ids";
 import {
-    Button,
     Card,
     CardActions,
     CardContent,
@@ -29,7 +27,6 @@ import {
     Typography,
 } from "@material-ui/core";
 import SupervisorAccountIcon from "@material-ui/icons/SupervisorAccount";
-import LiveHelpIcon from "@material-ui/icons/LiveHelp";
 import ErrorIcon from "@material-ui/icons/Error";
 import { trackIntercomEvent, IntercomEvents } from "common/intercom";
 
@@ -53,75 +50,6 @@ const useStyles = makeStyles((theme) => ({
         cursor: "pointer",
     },
 }));
-
-export function ClientInfo(props) {
-    const [browser, setBrowser] = useState();
-    const { t } = useTranslation("util");
-    useEffect(() => {
-        const doSetBrowser = async () => {
-            const Bowser = (await import("bowser")).default;
-            setBrowser(Bowser.getParser(window.navigator.userAgent));
-        };
-        doSetBrowser();
-    }, []);
-    const { baseId } = props;
-    const [userProfile] = useUserProfile();
-    if (!browser) {
-        return <GridLoading rows={5} />;
-    }
-    return (
-        <>
-            <GridLabelValue label={t("user")}>
-                <Typography id={build(baseId, ids.USER)}>
-                    {userProfile?.id}
-                </Typography>
-            </GridLabelValue>
-            <GridLabelValue label={t("browser")}>
-                <Typography id={build(baseId, ids.BROWSER)}>
-                    {browser.getBrowser().name} - {browser.getBrowser().version}
-                </Typography>
-            </GridLabelValue>
-            <GridLabelValue label={t("os")}>
-                <Typography id={build(baseId, ids.OS)}>
-                    {browser.getOS().name} - {browser.getOS().versionName} -
-                    {browser.getOS().version}
-                </Typography>
-            </GridLabelValue>
-            <GridLabelValue label={t("host")}>
-                <Typography
-                    id={build(baseId, ids.host)}
-                    className={constants.CHROMATIC_IGNORE}
-                >
-                    {window.location.origin}
-                </Typography>
-            </GridLabelValue>
-            <GridLabelValue label={t("timestamp")}>
-                <Typography
-                    id={build(baseId, ids.host)}
-                    className={constants.CHROMATIC_IGNORE}
-                >
-                    {new Date().toString()}
-                </Typography>
-            </GridLabelValue>
-        </>
-    );
-}
-
-function ContactSupport(props) {
-    const { baseId } = props;
-    const { t } = useTranslation("util");
-    return (
-        <Button
-            id={build(baseId, ids.CONTACT_SUPPORT_BUTTON)}
-            color="primary"
-            startIcon={<LiveHelpIcon />}
-            onClick={intercomShow}
-            style={{ marginLeft: "auto" }}
-        >
-            {t("contactSupport")}
-        </Button>
-    );
-}
 
 function ErrorHandler(props) {
     const { errorObject, baseId } = props;
